requirements.txt
aiohttp==3.9.5
asgiref==3.8.1
asyncio==3.4.3
better-profanity==0.7.0
fastapi==0.111.0
Flask==3.0.3
pymongo==4.7.3
python-telegram-bot==21.3
spacy==3.7.5
uvicorn==0.30.1
dnspython==2.6.1
bs4==0.0.2
python-dotenv==1.0.1
google-generativeai==0.7.0
langchain==0.2.5
langchain-community==0.2.5
logging==0.4.9.6
http==0.02